requirements.txt
Django==4.2.7
cyrtranslit==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1